[REF] l10n_it_ricevute_bancarie: Extract tests setup
