+++ l10n_it_ricevute_bancarie/tests/test_riba.py
@@ -4,11 +4,10 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import base64
-import os
 
 from odoo.exceptions import UserError
 from odoo.tests import Form
-from odoo.tools import config, safe_eval
+from odoo.tools import safe_eval
 
 from . import riba_common
 
@@ -93,82 +92,6 @@ class TestInvoiceDueCost(riba_common.TestRibaCommon):
         # ---- Set to Draft
         # Collection fees line has been unlink
         self.assertEqual(len(self.invoice.invoice_line_ids), 1)
-
-    def riba_sbf_common(self, configuration_id):
-        invoice = self._create_sbf_invoice()
-        invoice._onchange_riba_partner_bank_id()
-        invoice.action_post()
-        riba_move_line_id = False
-        for move_line in invoice.line_ids:
-            if move_line.account_id.id == self.account_rec1_id.id:
-                riba_move_line_id = move_line.id
-                line_ids = self.move_line_model.search(
-                    [
-                        "&",
-                        "|",
-                        ("riba", "=", "True"),
-                        ("unsolved_invoice_ids", "!=", False),
-                        ("account_id.internal_type", "=", "receivable"),
-                        ("reconciled", "=", False),
-                        ("distinta_line_ids", "=", False),
-                        ("move_id", "=", invoice.id),
-                    ]
-                )
-                self.assertEqual(len(line_ids), 1)
-                self.assertEqual(line_ids[0].id, move_line.id)
-        self.assertTrue(riba_move_line_id)
-
-        # issue wizard
-        wizard_riba_issue = self.env["riba.issue"].create(
-            {"configuration_id": configuration_id}
-        )
-        action = wizard_riba_issue.with_context(
-            {"active_ids": [riba_move_line_id]}
-        ).create_list()
-        riba_list_id = action and action["res_id"] or False
-        riba_list = self.distinta_model.browse(riba_list_id)
-        riba_list.confirm()
-        self.assertEqual(riba_list.state, "accepted")
-        self.assertEqual(invoice.state, "posted")
-
-        # Se la compute non viene invocata il test fallisce
-        riba_list._compute_acceptance_move_ids()
-        self.assertEqual(len(riba_list.acceptance_move_ids), 1)
-        self.assertEqual(len(riba_list.payment_ids), 0)
-
-        # I print the C/O distinta report
-        docargs = {
-            "doc_ids": riba_list.ids,
-            "doc_model": "riba.distinta",
-            "docs": self.env["riba.distinta"].browse(riba_list.ids),
-        }
-        data = self.env.ref("l10n_it_ricevute_bancarie.distinta_qweb")._render(docargs)
-        if config.get("test_report_directory"):
-            open(
-                os.path.join(config["test_report_directory"], "riba-list." + format),
-                "wb+",
-            ).write(data)
-
-        # credit wizard
-        wiz_accreditation = (
-            self.env["riba.accreditation"]
-            .with_context(
-                {
-                    "active_model": "riba.distinta",
-                    "active_ids": [riba_list_id],
-                    "active_id": riba_list_id,
-                }
-            )
-            .create(
-                {
-                    "bank_amount": 445,
-                    "expense_amount": 5,
-                }
-            )
-        )
-        wiz_accreditation.create_move()
-        self.assertEqual(riba_list.state, "accredited")
-        return invoice, riba_list
 
     def test_riba_sbf_maturation_flow(self):
         invoice, riba_list = self.riba_sbf_common(self.riba_config_sbf_maturation.id)
